Make `describe_successfully` available for all specs.
